requirements.txt
requests
python-dotenv
pyotp
qrcode 